+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## 1.0.7
+- Added logging if error occurs when fetching Battery max/min from database.
+
+## 1.0.6
+- Misconfiguration in mower database.
+
+## 1.0.5
+- Added better handling for unknowm mower models. Warning messages are raised in log file with model number.
+
+## 1.0.4
+- Corrected misspelling on getLastCompletedMow and related functions
+
 ## 1.0.3
 - Renamed getLastCutting to getLastCompletedMow
 - Renamed getNextCutting to getNextMow